requirements.txt
streamlit
PyPDF2
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
